+++ build_test/lib/src/stub_writer.dart
@@ -8,6 +8,8 @@ import 'package:build/build.dart';
 
 class StubAssetWriter implements AssetWriter {
   @override
-  Future writeAsString(Asset asset, {Encoding encoding: UTF8}) =>
-      new Future.value(null);
+  Future writeAsBytes(_, __) => new Future.value(null);
+
+  @override
+  Future writeAsString(_, __, {Encoding encoding}) => new Future.value(null);
 }